requirements.txt
annotated-types==0.7.0
anyio==4.5.2
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
flask==3.0.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.9.1
lxml==5.4.0
MarkupSafe==2.1.5
openai==1.82.0
packaging>=23.2,<25.0
pillow==10.4.0
pydantic==2.10.6
pydantic-core==2.27.2
pydub==0.25.1
PyMuPDF==1.24.11
pytesseract==0.3.13
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-telegram-bot==21.6
requests==2.32.3
sniffio==1.3.1
SpeechRecognition==3.10.4
tqdm==4.67.1
typing-extensions==4.13.2
urllib3==2.2.3
websocket-client==1.8.0
werkzeug==3.0.6
zipp==3.20.2

# Nuevas dependencias para arquitectura refactorizada
pydantic-settings==2.1.0
chromadb==0.4.22
tiktoken==0.5.2
langchain-text-splitters==0.2.0
